Move add inactive tasks toggle to menubar
